add dataset.py

+++ dataset.py
@@ -16,11 +16,12 @@ def main():
         os.makedirs(path, exist_ok=True)
         print('tar -xf ./Dataset/ILSVRC2012_img_{}.tar -C {}'.format(phase, path))
         os.system('tar -xf ./Dataset/ILSVRC2012_img_{}.tar -C {}'.format(phase, path))
-        for tar in os.listdir(path):
-            print('tar -xf {}/{} -C {}/{}'.format(path, tar, path, tar.split('.')[0]))
-            os.makedirs('{}/{}'.format(path, tar.split('.')[0]), exist_ok=True)
-            os.system('tar -xf {}/{} -C {}/{}'.format(path, tar, path, tar.split('.')[0]))
-            os.remove('{}/{}'.format(path, tar))
+        if phase == 'train':
+            for tar in os.listdir(path):
+                print('tar -xf {}/{} -C {}/{}'.format(path, tar, path, tar.split('.')[0]))
+                os.makedirs('{}/{}'.format(path, tar.split('.')[0]), exist_ok=True)
+                os.system('tar -xf {}/{} -C {}/{}'.format(path, tar, path, tar.split('.')[0]))
+                os.remove('{}/{}'.format(path, tar))
 
 
 if __name__ == '__main__':
